+++ src/pages/instructorRegister/InstructorRegister.tsx
@@ -1,13 +1,5 @@
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useQueryClient } from '@tanstack/react-query';
-import { useEffect, useState } from 'react';
-import { FormProvider, useController, useForm } from 'react-hook-form';
-import { useNavigate } from 'react-router-dom';
-import {
-  useGetInstructorRegisterInfo,
-  usePatchInstructorRegisterInfo,
-  usePostInstructor,
-} from '@/pages/instructorRegister/apis/queries';
+import { FormProvider, useController } from 'react-hook-form';
+import { useGetInstructorRegisterInfo } from '@/pages/instructorRegister/apis/queries';
 import CareerSection from '@/pages/instructorRegister/components/CareerSection/CareerSection';
 import DancerNameSection from '@/pages/instructorRegister/components/DancerNameSection/DancerNameSection';
 import ImageUploadSection from '@/pages/instructorRegister/components/ImageUploadSection/ImageUploadSection';
@@ -15,78 +7,31 @@ import IntroductionSection from '@/pages/instructorRegister/components/Introduct
 import PersonalSNSSection from '@/pages/instructorRegister/components/PersonalSNSSection/PersonalSNSSection';
 import VideoLinkSection from '@/pages/instructorRegister/components/VideoLinkSection/VideoLinkSection';
 import * as styles from '@/pages/instructorRegister/instructorRegister.css';
-import { instructorRegisterSchema } from '@/pages/instructorRegister/schema/instructorRegisterSchema';
-import type {
-  duplicateStateTypes,
-  instructorRegisterFormTypes,
-} from '@/pages/instructorRegister/types/instructorRegisterForm';
-import { ROUTES_CONFIG } from '@/routes/routesConfig';
 import { useGetRole } from '@/shared/apis/queries';
 import BoxButton from '@/shared/components/BoxButton/BoxButton';
 import Divider from '@/shared/components/Divider/Divider';
 import Head from '@/shared/components/Head/Head';
-import { notify } from '@/shared/components/Toast/Toast';
-import { authKeys, teacherKeys } from '@/shared/constants/queryKey';
-import { USER_ROLE } from '@/shared/constants/userRole';
 import useImageUploader from '@/shared/hooks/useImageUploader';
-import { setAccessToken, setRefreshToken } from '@/shared/utils/handleToken';
+import useInstructorRegisterForm from './hooks/useInstructorRegisterForm';
 
 const InstructorRegister = () => {
-  const [duplicateState, setDuplicateState] = useState<duplicateStateTypes>(null);
-
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
-
-  // 강사 등록
-  const { mutate: instructorRegisterMutate } = usePostInstructor();
-
-  // 강사 수정
   const { data } = useGetRole();
   const userRole = data?.role;
 
-  const { mutate: instructorPatchMutate } = usePatchInstructorRegisterInfo();
+  // 강사 수정용 이전 정보 조회
   const { data: prevInstructorData } = useGetInstructorRegisterInfo(userRole ?? '');
 
-  const methods = useForm({
-    resolver: zodResolver(instructorRegisterSchema),
-    mode: 'onTouched',
-    defaultValues: {
-      nickname: '',
-      detail: '',
-      instagram: '',
-      youtube: '',
-      imageUrls: '',
-      educations: [''],
-      experiences: [''],
-      prizes: [''],
-      videoUrls: [''],
-      isEduNoneChecked: false,
-      isCareerNoneChecked: false,
-      isPrizeNoneChecked: false,
-      isVideoNoneChecked: false,
-    },
-  });
+  const { methods, handleSubmit, isEditMode, isButtonActive, duplicateState, setDuplicateState } =
+    useInstructorRegisterForm({ userRole, prevInstructorData });
 
-  const {
-    reset,
-    control,
-    formState: { isDirty, isValid },
-    handleSubmit,
-  } = methods;
-
-  const { field } = useController({
+  const { field: imageField } = useController({
     name: 'imageUrls',
-    control,
+    control: methods.control,
   });
-
-  const isEditMode = userRole === USER_ROLE.TEACHER;
-  const isButtonActive = isEditMode
-    ? isDirty && isValid && duplicateState === 'available'
-    : isValid && duplicateState === 'available';
 
   // 이미지 업로드 관련
   const handleImageUploadSuccess = (url: string) => {
-    field.onChange(url);
+    imageField.onChange(url);
   };
 
   const { previewImg, imgRef, handleUploaderClick, uploadImgFile } = useImageUploader(
@@ -95,95 +40,9 @@ const InstructorRegister = () => {
     prevInstructorData?.profileImage
   );
 
-  // form submit 함수
-  const onSubmit = (data: instructorRegisterFormTypes) => {
-    const updatedInfo = {
-      nickname: data.nickname.trim(),
-      imageUrls: [data.imageUrls],
-      detail: data.detail.trim(),
-
-      instagram: data.instagram?.trim() === '' ? null : `${data.instagram?.trim()}`,
-      youtube: data.youtube?.trim() === '' ? null : `${data.youtube?.trim()}`,
-
-      educations: data.isEduNoneChecked ? [] : data.educations.filter((education) => education.trim() !== ''),
-      experiences: data.isCareerNoneChecked ? [] : data.experiences.filter((experience) => experience.trim() !== ''),
-      prizes: data.isPrizeNoneChecked ? [] : data.prizes.filter((prize) => prize.trim() !== ''),
-
-      videoUrls: data.isVideoNoneChecked ? [] : data.videoUrls.filter((url) => url.trim() !== ''),
-    };
-
-    // 강사 등록 성공 함수
-    const onSuccessRegister = (response: { data: { accessToken: string; refreshToken: string } }) => {
-      const { accessToken, refreshToken } = response.data;
-
-      setAccessToken(accessToken);
-      setRefreshToken(refreshToken);
-
-      queryClient.invalidateQueries({ queryKey: authKeys.role.queryKey });
-
-      navigate(ROUTES_CONFIG.instructorRegisterCompletion.path);
-    };
-
-    // 강사 수정 성공 함수
-    const onSuccessEdit = () => {
-      // 강사 수정용 조회 API invalidate
-      queryClient.invalidateQueries({ queryKey: teacherKeys.me.queryKey });
-
-      // 댄서 조회 API invalidate
-      queryClient.invalidateQueries({ queryKey: teacherKeys.list.queryKey });
-
-      navigate(ROUTES_CONFIG.mypage.withTab('teacher'));
-      notify({ message: '수정이 완료되었어요.', icon: 'success' });
-    };
-
-    const onError = () => {
-      navigate(ROUTES_CONFIG.error.path);
-    };
-
-    if (isEditMode) {
-      // 강사 수정 (이미 강사 등록된 경우)
-      instructorPatchMutate(updatedInfo, {
-        onSuccess: onSuccessEdit,
-        onError,
-      });
-    } else {
-      // 강사 등록 (아직 강사 등록 안된 경우)
-      instructorRegisterMutate(updatedInfo, {
-        onSuccess: onSuccessRegister,
-        onError,
-      });
-    }
-  };
-
-  useEffect(() => {
-    if (!prevInstructorData) return;
-
-    // TODO: prev data에 dancerName 추가되면 주석 제거
-    // if (prevInstructorData.dancerName) {
-    // setDuplicateState('available');
-    // }
-
-    reset({
-      nickname: prevInstructorData.nickname || 'd',
-      imageUrls: prevInstructorData.profileImage || '',
-      instagram: prevInstructorData.instagram || '',
-      youtube: prevInstructorData.youtube || '',
-      educations: prevInstructorData.educations?.length ? prevInstructorData.educations : [''],
-      experiences: prevInstructorData.experiences?.length ? prevInstructorData.experiences : [''],
-      prizes: prevInstructorData.prizes?.length ? prevInstructorData.prizes : [''],
-      detail: prevInstructorData.detail || '',
-      videoUrls: prevInstructorData.videoUrls?.length ? prevInstructorData.videoUrls : [''],
-
-      isEduNoneChecked: !(prevInstructorData.educations?.length > 0),
-      isCareerNoneChecked: !(prevInstructorData.experiences?.length > 0),
-      isPrizeNoneChecked: !(prevInstructorData.prizes?.length > 0),
-      isVideoNoneChecked: !(prevInstructorData.videoUrls?.length > 0),
-    });
-  }, [prevInstructorData, reset]);
-
   return (
     <FormProvider {...methods}>
-      <form onSubmit={handleSubmit(onSubmit)}>
+      <form onSubmit={handleSubmit}>
         <div className={styles.containerStyle}>
           <div className={styles.sectionWrapperStyle}>
             <div className={styles.titleStyle}>
